Refactor Namespace to better match usage of model
Since these models will be used to map db tables, it makes I'm turn between naming it's namespace.
I contemplated between ViewModel, POCO, & DTO. DTO seems to make more sense at the moment.

Might change later
