+++ app/wama.core/Models/Service/ICheckInService.cs
@@ -1,4 +1,4 @@
-﻿using WAMA.Core.Models.POCOs;
+﻿using WAMA.Core.Models.DTOs;
 
 namespace WAMA.Core.Models.Service
 {
