+++ app/wama.core/Models/Service/IUserAccountService.cs
@@ -1,5 +1,5 @@
 ﻿using System.Threading.Tasks;
-using WAMA.Core.Models.POCOs;
+using WAMA.Core.Models.DTOs;
 
 namespace WAMA.Core.Models.Service
 {
