+++ app/wama.core/Services/UserAccountService.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 using WAMA.Core.Models;
-using WAMA.Core.Models.POCOs;
+using WAMA.Core.Models.DTOs;
 using WAMA.Core.Models.Service;
 
 namespace WAMA.Core.Services
